requirements.txt
fastapi
uvicorn
transformers
torch
pytesseract
pillow
textblob
python-docx
pymupdf
pydantic
python-multipart
numpy
scikit-learn